Use american english spelling of favorites
